+++ CHANGELOG.md
@@ -440,11 +440,12 @@
 * [X] 移除`breeze-serializer`
 * [X] breeze-serialization 添加`MapSerializer`
 * [X] breeze-serialization 完善`MapSerializer`
-* [X] breeze-serialization 为某一数据类型的`Serializer`如`JsonSerializer`提供作为伴生对象的默认实现。
+* [X] breeze-serialization 为某一数据类型的`Serializer`如`JsonSerializer`提供作为伴生对象的默认实现
 * [X] breeze-core 将`Any.expand`改为`List<T>.expand`
 * [X] 移除`breeze-mapper`，`breeze-serializer`，`breeze-generator`并合并到`breeze-unstable`
 * [X] 添加`List<T>.pin()`扩展
 * [X] 整理目录
+* [X] 添加`buildText`和`Iterable.select()`
 * [X] breeze-serialization 重构代码
 * [ ] breeze-dsl 重构代码
   * [X] breeze-dsl-xml